[#24] UIImage+Extension downsample 함수 분리
